+++ src/index.ts
@@ -1,1 +1,8 @@
 export { Globe } from "./components/Globe";
+export type {
+  GlobeConfiguration,
+  GlobeControls,
+  GlobeProps,
+  Trajectory,
+  WMSConfig,
+} from "./components/Globe";