+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Globe Changelog
 
+## 1.3.0 [2025-10-14]
+
+- Updated Globe component to better import into strangely styled parents
+
+## 1.2.0 [2025-10-13]
+
+- Removed Resium
+
 ## 1.1.1 [2025-10-13]
 
 - Updated Globe component to isolate internal styling